+++ rbc/shader/src/geometry/set_submesh.cpp
@@ -3,8 +3,8 @@
 using namespace luisa::shader;
 
 [[kernel_1d(128)]] int kernel(
-	Buffer<uint>& dst,
+	Buffer<uint16>& dst,
 	Buffer<uint>& offsets) {
-	dst.write(offsets.read(kernel_id()) + dispatch_id().x, kernel_id());
+	dst.write(offsets.read(kernel_id()) + dispatch_id().x, (uint16)kernel_id());
 	return 0;
 }
